requirements.txt
google-genai>=1.0.0
datasets>=2.21.0
transformers>=4.45.0
trl>=0.11.0
peft>=0.13.0
bitsandbytes>=0.43.0
accelerate>=0.34.0
torch>=2.2.0
python-dotenv>=1.0.1
